Player parent to moving vehicle successful
Successful parenting of player to moving vehicle without any kind of visual stuttering.
It works by having two different Player Controller scripts, one is the Unity default FirstPersonController.cs and this one is activated while the player is outside the vehicle, and the other is a custom player controller script (PlayerUltimateController.cs) that handles the movement while inside the vehicle (without using the character controller component), during all this there is a third script called PlayerParenting.cs that is handling the toggling on and off of both player controller scripts.
Quite a mess, but hey IT WORKS! (for now)